# PERF: compare RangeIndex to equal RangeIndex