Fastpath for RangeIndexes with equal `_range` attribute.

Notice this works even if the RangeIndexes are different.

Example:

```python
>>>  n = 100_000
>>> rng1 = pd.RangeIndex(n)
>>> rng2 = pd.RangeIndex(n)
>>> %timeit rng1 == rng2
374 µs ± 15.1 µs per loop  # master
10.4 µs ± 157 ns per loop  # this PR
```

Somewhat related to #37109